+++ TShirt/analog.h
@@ -1,4 +1,4 @@
 /** Definitions for the analog converter **/
 
 void ad_init(unsigned char channel);
-uint8_t ad_sample(void);
+unsigned int ad_sample(void);
